chore: repository 파일 통합
