--- domain/repositories/ShareRepository.ts
@@ -1,11 +1,0 @@
-import type { Prisma, Share } from "@/prisma/generated";
-
-export interface FindByOwnerAndStatus {
-    where: Prisma.ShareWhereInput,
-    offset: number,
-    itemsPerPage: number,
-}
-
-export interface ShareRepository {
-  findByOwnerAndStatus(shares: FindByOwnerAndStatus): Promise<(Partial<Share> & {thumbnailUrl: string | null})[]>;
-}
